Se crea la DataBase

+++ db/migrations/20190406193155_create_producto.js
@@ -4,7 +4,7 @@ exports.up = function(knex, Promise) {
         knex.schema.createTable('producto', (table) => {
           table.increments('id').primary().unsigned();
           table.string('nombre');
-          table.double('precio');
+          table.decimal('precio');
           table.integer('stock');
           table.string('referencia').unique();
         })
